Fixed some integer conversion warning

+++ testrunner.h
@@ -45,7 +45,7 @@ extern u32 musashiFC;
 inline u8 get8(u8 *p, u32 addr) {
     return p[addr & 0xFFFF]; }
 inline u16 get16(u8 *p, u32 addr) {
-    return get8(p, addr) << 8 | get8(p, addr + 1); }
+    return (u16)(get8(p, addr) << 8 | get8(p, addr + 1)); }
 inline void set8(u8 *p, u32 addr, u8 val) {
     p[addr & 0xFFFF] = val; }
 inline void set16(u8 *p, u32 addr, u16 val) {
